+++ tests/multi_span_tests.cpp
@@ -15,7 +15,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #include <UnitTest++/UnitTest++.h>
-#include <multi_span.h>
+#include <gsl/multi_span.h>
 
 #include <iostream>
 #include <list>
